setup.py: add dependencies to install_requires; use setuptools.find_packages; use pytest-runner; add trove classifiers for PyPI
The inlined test suite is now installed alongside the package.
If we prefer we can decide to move the 'test' folder outside of the package, and not install it.
Pytest can discover it without problems.

TODO(anthrotype): comment out long_description once we the repo has a README file which we can use