mathFunctions: add workaround for inconsistent round() behavior in Python < 3.6
On Python 3, FontTools' `round3` function is simply an alias of the built-in
`round()` function.

The latter has inconsistent behaviour when the first argument is a float or an
int, and when the second argument (ndigits) is passed with value None.

floats accept a second argument ndigits=None just fine:

>>> round(2.0, None)
2

Whereas integers raise a TypeError if the second argument is None:

>>> round(2, None)
TypeError: 'NoneType' object cannot be interpreted as an integer

This bug has been fixed in the upcoming Python 3.6
For more info, see: https://bugs.python.org/issue27936