mathFunctions.py: don't wrap fonttools `round3` when running on py27 as the pure-python backport already accepts a second None argument
It is only the built-in round() in Python 3.x (< 3.6) that has that inconsistent behavior (fonttols round3 is an alias of the built-in round in Python 3), not the
pure-python round3 backport used on Python 2.7.

Note that the current fonttools HEAD contains the same workaround:
https://github.com/fonttools/fonttools/pull/757

We don't want to duplicate it here, so we shall drop this and update fontMath's
minimum fonttools requirement as soon as a new version of fonttools is out, which
is very soon.